move shared logic to src/git-worktree-utils.ts
