+++ src/git-worktree-add.ts
@@ -2,108 +2,11 @@
 
 import { join } from "node:path";
 import { existsSync } from "node:fs";
-import { readFile, readdir } from "node:fs/promises";
+import { readdir } from "node:fs/promises";
 import { $ } from "zx";
-import * as yaml from "yaml";
+import { findGitRoot, getMainBranch, branchExists } from "./git-worktree-utils.js";
 
 $.verbose = false;
-
-interface GitWorktreeConfig {
-  repositoryUrl: string;
-  mainBranch: string;
-  createdAt: string;
-}
-
-async function getMainBranch(gitRoot?: string): Promise<string> {
-  try {
-    // First, try to find the config file in current directory or project root
-    let configPath = join(process.cwd(), "git-worktree-config.yaml");
-
-    // If not found in current directory and we have gitRoot, try looking in the parent of gitRoot
-    if (!existsSync(configPath) && gitRoot) {
-      const projectRoot = join(gitRoot, "..");
-      configPath = join(projectRoot, "git-worktree-config.yaml");
-    }
-
-    if (existsSync(configPath)) {
-      const configContent = await readFile(configPath, "utf-8");
-      const config: GitWorktreeConfig = yaml.parse(configContent);
-      return config.mainBranch;
-    }
-
-    // If we have a git root, try git commands from there
-    if (gitRoot) {
-      try {
-        const symbolicRef =
-          await $`cd ${gitRoot} && git symbolic-ref refs/remotes/origin/HEAD 2>/dev/null`.text();
-        const match = symbolicRef.match(/refs\/remotes\/origin\/(.+)/);
-        if (match) {
-          return match[1].trim();
-        }
-      } catch {
-        // Try alternative method
-        try {
-          const remoteBranches =
-            await $`cd ${gitRoot} && git branch -r --points-at refs/remotes/origin/HEAD 2>/dev/null`.text();
-          if (remoteBranches.trim()) {
-            // Parse output like "  origin/dev -> origin/dev" or "  origin/dev"
-            const match = remoteBranches.match(/origin\/([^\s\-]+)/);
-            if (match) {
-              return match[1].trim();
-            }
-          }
-        } catch {
-          // Continue to fallback
-        }
-      }
-    }
-
-    // Default fallback
-    return "main";
-  } catch {
-    return "main";
-  }
-}
-
-async function branchExists(
-  branchName: string,
-  gitRoot?: string
-): Promise<{ exists: boolean; local: boolean; remote: boolean }> {
-  try {
-    let localBranches: string;
-    let remoteBranches: string;
-
-    if (gitRoot) {
-      // Execute commands from the git root directory
-      localBranches = await $`cd ${gitRoot} && git branch --list ${branchName} 2>/dev/null`.text();
-      remoteBranches = await $`cd ${gitRoot} && git branch -r --list "origin/${branchName}" 2>/dev/null`.text();
-    } else {
-      // Execute commands from current directory
-      localBranches = await $`git branch --list ${branchName} 2>/dev/null`.text();
-      remoteBranches = await $`git branch -r --list "origin/${branchName}" 2>/dev/null`.text();
-    }
-
-    const localExists = localBranches.trim().length > 0;
-    const remoteExists = remoteBranches.trim().length > 0;
-
-    return {
-      exists: localExists || remoteExists,
-      local: localExists,
-      remote: remoteExists,
-    };
-  } catch {
-    return { exists: false, local: false, remote: false };
-  }
-}
-
-async function findGitRoot(): Promise<string | null> {
-  try {
-    const gitRoot = await $`git rev-parse --show-toplevel 2>/dev/null`.text();
-    return gitRoot.trim();
-  } catch {
-    return null;
-  }
-}
 
 async function gwtadd(folderName: string) {
   if (!folderName) {
