+++ src/git-worktree-completion.ts
@@ -1,137 +1,6 @@
 #!/usr/bin/env tsx
 
-import { join } from "node:path";
-import { existsSync } from "node:fs";
-import { readdir } from "node:fs/promises";
-import { $ } from "zx";
-
-$.verbose = false;
-
-interface Worktree {
-	path: string;
-	HEAD: string;
-	branch: string;
-	bare?: boolean;
-}
-
-async function findWorktreesFromExisting(rootPath: string): Promise<Worktree[]> {
-	// Find any existing worktree to use for git commands
-	const entries = await readdir(rootPath, { withFileTypes: true });
-	let foundWorktree: string | null = null;
-	
-	// First try direct subdirectories
-	for (const entry of entries) {
-		if (entry.isDirectory()) {
-			const dirPath = join(rootPath, entry.name);
-			const gitPath = join(dirPath, ".git");
-			if (existsSync(gitPath)) {
-				foundWorktree = dirPath;
-				break;
-			}
-		}
-	}
-	
-	// If not found in direct subdirectories, try one level deeper
-	if (!foundWorktree) {
-		for (const entry of entries) {
-			if (entry.isDirectory()) {
-				const subDir = join(rootPath, entry.name);
-				try {
-					const subEntries = await readdir(subDir, { withFileTypes: true });
-					for (const subEntry of subEntries) {
-						if (subEntry.isDirectory()) {
-							const dirPath = join(subDir, subEntry.name);
-							const gitPath = join(dirPath, ".git");
-							if (existsSync(gitPath)) {
-								foundWorktree = dirPath;
-								break;
-							}
-						}
-					}
-					if (foundWorktree) break;
-				} catch {
-					// Skip if we can't read subdirectory
-				}
-			}
-		}
-	}
-	
-	if (!foundWorktree) {
-		return [];
-	}
-	
-	// Use the found worktree to get the complete list via git
-	try {
-		const output = await $`cd ${foundWorktree} && git worktree list --porcelain`.text();
-		const worktrees: Worktree[] = [];
-		let currentWorktree: Partial<Worktree> = {};
-		
-		for (const line of output.split('\n')) {
-			if (line.startsWith('worktree ')) {
-				if (currentWorktree.path) {
-					worktrees.push(currentWorktree as Worktree);
-				}
-				currentWorktree = { path: line.slice(9) };
-			} else if (line.startsWith('HEAD ')) {
-				currentWorktree.HEAD = line.slice(5);
-			} else if (line.startsWith('branch ')) {
-				currentWorktree.branch = line.slice(7);
-			} else if (line === 'bare') {
-				currentWorktree.bare = true;
-			}
-		}
-		
-		if (currentWorktree.path) {
-			worktrees.push(currentWorktree as Worktree);
-		}
-		
-		return worktrees;
-	} catch {
-		return [];
-	}
-}
-
-async function getAllWorktrees(): Promise<Worktree[]> {
-	try {
-		let worktrees: Worktree[] = [];
-		
-		// Check if we're in a git repository
-		try {
-			const output = await $`git worktree list --porcelain 2>/dev/null`.text();
-			let currentWorktree: Partial<Worktree> = {};
-			
-			for (const line of output.split('\n')) {
-				if (line.startsWith('worktree ')) {
-					if (currentWorktree.path) {
-						worktrees.push(currentWorktree as Worktree);
-					}
-					currentWorktree = { path: line.slice(9) };
-				} else if (line.startsWith('HEAD ')) {
-					currentWorktree.HEAD = line.slice(5);
-				} else if (line.startsWith('branch ')) {
-					currentWorktree.branch = line.slice(7);
-				} else if (line === 'bare') {
-					currentWorktree.bare = true;
-				}
-			}
-			
-			if (currentWorktree.path) {
-				worktrees.push(currentWorktree as Worktree);
-			}
-		} catch {
-			// Not in a git repository, check if we're in a project root
-			const configPath = join(process.cwd(), "git-worktree-config.yaml");
-			if (existsSync(configPath)) {
-				// We're in a project root, scan for worktrees
-				worktrees = await findWorktreesFromExisting(process.cwd());
-			}
-		}
-		
-		return worktrees;
-	} catch {
-		return [];
-	}
-}
+import { getAllWorktrees, Worktree, cleanBranchName } from "./git-worktree-utils.js";
 
 async function getAvailableBranches(command: string): Promise<string[]> {
 	const worktrees = await getAllWorktrees();
@@ -143,7 +12,7 @@ async function getAvailableBranches(command: string): Promise<string[]> {
 	if (command === "gwtswitch" || command === "gwtremove") {
 		// For gwtswitch and gwtremove, return existing worktree branches
 		return worktrees
-			.map(wt => wt.branch ? wt.branch.replace('refs/heads/', '') : '')
+			.map(wt => cleanBranchName(wt.branch))
 			.filter(branch => branch.length > 0);
 	} else if (command === "gwtadd") {
 		// For gwtadd, return available remote branches (not implemented yet)
